platform/pyproject.toml
[project]
name = "muppet-platform"
version = "0.1.0"
description = "Internal developer platform for creating and managing backend applications"
authors = [
    {name = "Muppet Platform Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "PyGithub>=2.1.1",
    "python-json-logger>=2.0.7",
    "mcp>=1.0.0",
    "hypothesis==6.92.1",
    "aiohttp==3.9.1",
    "aiofiles>=23.2.1",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "moto==4.2.14",
]

[project.scripts]
mcp-server = "src.platform_mcp.cli:run_mcp_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "autopep8>=2.2.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "moto>=4.2.14",
    "mypy>=1.7.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
]
